:white_check_mark: feat(get job): add tests for get job command
update get job to be used with the entity framework

+++ resume builder/cli/commands/search/SearchJobCommand.cs
@@ -2,29 +2,28 @@ using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 using resume_builder.cli.settings;
 using resume_builder.models;
-using resume_builder.models;
 using Spectre.Console;
 using Spectre.Console.Cli;
 
 namespace resume_builder.cli.commands.search;
 
-public class SearchJobCommand : JobOutputCommand<SearchJobSettings>
+public class SearchJobCommand: JobOutputCommand<SearchJobSettings>
 {
     public override int Execute([NotNull] CommandContext context, [NotNull] SearchJobSettings settings)
     {
-        Dictionary<long, Job> rows;
+        Dictionary<int, Job> rows;
 
         try
         {
             ResumeContext database = new();
             rows = database.Jobs.AsQueryable()
-                .Where(job => settings.Terms.Any(term =>
-                job.Description.Contains(term, StringComparison.OrdinalIgnoreCase) ||
-                job.Experience.Contains(term, StringComparison.OrdinalIgnoreCase)
-                // || job.Skills.Any(skill => skill.Name.Contains(term, StringComparison.OrdinalIgnoreCase))
-                ))
-                .Select((job, i) => new KeyValuePair<long, Job>(i, job))
-                .ToDictionary(tuple => tuple.Key, tuple => tuple.Value);
+                           .Where(job => settings.Terms.Any(term =>
+                                    job.Description.Contains(term, StringComparison.OrdinalIgnoreCase) ||
+                                    job.Experience.Contains(term, StringComparison.OrdinalIgnoreCase)
+                                // || job.Skills.Any(skill => skill.Name.Contains(term, StringComparison.OrdinalIgnoreCase))
+                            ))
+                           .Select((job, i) => new KeyValuePair<int, Job>(i, job))
+                           .ToDictionary(tuple => tuple.Key, tuple => tuple.Value);
         }
         catch(Exception e)
         {
@@ -48,7 +47,7 @@ public class SearchJobCommand : JobOutputCommand<SearchJobSettings>
     }
 }
 
-public class SearchJobSettings : JobOutputSettings
+public class SearchJobSettings: JobOutputSettings
 {
     [CommandArgument(0, "[terms]")] //[CommandOption("-t|--terms")]
     [Description("search terms to search for in the job's description, experience, or skills")]
